fix(rust-consensus): address RUB-497 review — 32-bit witness_len guard, config panic→error, doc/coverage
Follow-up to PR #2187 review (Copilot + Codex):

- tx.rs (P1, fork-class): guard witness_len with isize::MAX instead of i64::MAX.
  isize::MAX == Go math.MaxInt on every target (2^63-1 / 2^31-1), so the guard
  stays byte-identical to Go AND bounds the value at/under usize::MAX, preventing
  a 32-bit `as usize` truncation from accepting a witness_len Go rejects.

- suite_registry / genesis / cli (P2): the with_suites structural-range assert
  mirrors Go's trusted-caller panic, but the config/CLI paths never pre-rejected
  structural ids, so untrusted genesis/CLI JSON could panic instead of returning
  "bad suite_registry" (DoS). Mirror Go's config pre-check: reject 0xF0..=0xFE in
  genesis validate_suite_registry_item and cli build_suite_registry_from_json.
  Re-export is_structural_witness_carrier_suite_id from the crate root.

- fuzz: mask structural ids out of the custom-suite range so with_suites' assert
  is not tripped as a false crash.

- suite_registry.rs (Copilot): move is_structural_witness_carrier_suite_id above
  the SuiteRegistry doc comment so the struct keeps its own docs.

- tests/coverage: witness_len-overflow parse test; genesis + cli regression tests
  that a structural id returns Err (not panic); drop unreachable compact_size
  branches in the test helper.

Conformance-vector coverage (Codex P1) is deferred to a cross-client follow-up:
the merged Go reference (RUB-496) added no 0xF0 CV vectors, so adding them here
would exceed the Rust mirror and break this PR's parity-exempt basis.

+++ clients/rust/crates/rubin-consensus/src/tx.rs
@@ -405,8 +405,8 @@ fn validate_simplicity_envelope_signature(signature: &[u8]) -> Result<(), TxErro
     }
     r.read_bytes(program_len_u64 as usize)?;
     let (witness_len_u64, _) = read_compact_size(&mut r)?;
-    // Go rejects witness_len > math.MaxInt (== i64::MAX on the 64-bit consensus target).
-    if witness_len_u64 > i64::MAX as u64 {
+    // `isize::MAX` == Go's `math.MaxInt` on every target: byte-identical, and blocks a 32-bit `as usize` truncation.
+    if witness_len_u64 > isize::MAX as u64 {
         return Err(TxError::new(
             ErrorCode::TxErrParse,
             "Simplicity witness_len overflows int",
